socket-bridge: prints packets

+++ tools/socket-bridge/NetworkLocalServer.cpp
@@ -1,4 +1,5 @@
 #include "NetworkLocalServer.h"
+#include "ConsoleColors.h"
 #include <QDebug>
 #include <QFile>
 
@@ -6,12 +7,29 @@ NetworkLocalServer::NetworkLocalServer(QObject *parent) : QLocalServer(parent) {
     connect(this, &QLocalServer::newConnection, this, &NetworkLocalServer::newConnectionAvailable);
 }
 
+NetworkLocalServer::NetworkLocalServer(const QString &objectName, QObject *parent) : NetworkLocalServer(parent) {
+    setObjectName(objectName);
+}
+
+void NetworkLocalServer::print(QString str, const QByteArray &data) {
+    auto dbg = qDebug().noquote().nospace();
+    QString prefix = QString("%1 %2:").arg(str).arg(data.size());
+    prefix += QString(30-prefix.count(),' ');
+    auto len = data.left(4).toHex(' ') + " ";
+    auto id = data.mid(4,4).toHex(' ') + " ";
+    auto other = data.mid(8,8).toHex(' ');
+    dbg << prefix << cYLW  << len << cORG << id << cRST << other;
+}
+
+
 void NetworkLocalServer::writeData(QByteArray data) {
+    print(objectName()+" Write", data);
+    qDebug() << "-";
     if (socket_) {
         socket_->write(data);
         socket_->flush();
     } else {
-        qDebug() << serverName() << " Socket is close, sink data";
+        qDebug() << objectName() << " Socket is close, sink data";
     }
 }
 
@@ -24,22 +42,23 @@ void NetworkLocalServer::deleteSocket() {
 }
 
 void NetworkLocalServer::stop() {
-    if (socket_) socket_->close();
+    if (socket_) {
+        socket_->flush();
+        socket_->close();
+    }
     deleteSocket();
     close();
 }
 
 void NetworkLocalServer::start(const QString &port) {
-    if (QFile::exists(port)) {
-        QFile::remove(port);
-    }
+    if (QFile::exists(port))    QFile::remove(port);
     listen(port);
-    qDebug() << serverName() << " is listening";
+    qDebug() << objectName() << " is listening on:" << serverName();
 }
 
 void NetworkLocalServer::newConnectionAvailable() {
     if (socket_) {
-        qDebug() << serverName() << "Sorry I can accept only one connection";
+        qDebug() << objectName() << "Sorry, I can accept only one connection";
         return;
     }
     socket_ = nextPendingConnection();
@@ -49,7 +68,7 @@ void NetworkLocalServer::newConnectionAvailable() {
 }
 
 void NetworkLocalServer::onSocketErrorOccurred(QLocalSocket::LocalSocketError socketError) {  Q_UNUSED(socketError)
-    qDebug() << serverName() << "Socket error: " << socket_->errorString();
+    qDebug() << objectName() << "Socket error: " << socket_->errorString();
     deleteSocket();
 }
 
@@ -63,5 +82,7 @@ void NetworkLocalServer::onSocketStateChanged(QLocalSocket::LocalSocketState soc
 }
 
 void NetworkLocalServer::onSocketData() {
-    emit dataReady(socket_->readAll());
+    auto data = socket_->readAll();
+    print(objectName()+" Read",data );
+    emit dataReady(data);
 }
